Initial commit
Fully vectorized neural network implementation for MATLAB and Octave. Takes in an arbitrary topology vector of the form [input L1 L2 ... output] and builds a series of fully connected layers. Activation functions are specified in a cell vector of length L-1, where L is the length of the topology vector. An error function is passed to the training function to generalize the implementation.

A sample AND gate implementation is provided.